Process commit
